+++ lib/handlebars-compiler.ts
@@ -1,165 +1,80 @@
-import path from 'path';
-import fs from 'fs';
+import * as ParsedOptions from './types/plugin-options/parsed';
 
+import path from 'path';
 import Handlebars from 'handlebars';
 
-import type { SourceData, CompiledData, Helper } from './types'
-import { HandlebarsPluginOptions, TemplateSpecification } from './types';
+import { CompileResult, TemplateSpecification } from './types/handlebars';
+
+type CompiledData = [string, TemplateSpecification]
 
 export default class HandlebarsCompiler {
-    handlebarsPluginOptions: HandlebarsPluginOptions
-    cache: Map<string, string>
+    compileOptions: CompileOptions
+    partials: ParsedOptions.Partials
+    helpers: ParsedOptions.Helpers
+    templateData: ParsedOptions.TemplateData
 
-    constructor(handlebarsPluginOptions: HandlebarsPluginOptions) {
-        this.handlebarsPluginOptions = handlebarsPluginOptions
-        this.cache = new Map();
+    constructor(parsedOptions: ParsedOptions.IParsedOptions) {
+        Object.entries(parsedOptions).forEach(([key, value]) => {
+            this[key] = value
+        })
     }
 
-    static renamePartial(source: string, partialName: string, newPartialName: string) {
-        source = source.replaceAll(new RegExp(`(\\{\\{>(\\n|\\s)*)(${partialName})`, 'g'), `$1${newPartialName} `)
-        return source
-    }
-
-	static getPartialSource(partialPath: string, genesisFileDirectory: string, currentFile: string): string | undefined {
-		const relativeFileDirectory = path.dirname(currentFile)
-		const relativePartialPath = path.join(relativeFileDirectory, partialPath)
-		const partialAbsolutePath = path.resolve(genesisFileDirectory, `${relativePartialPath}.hbs`)
-		let partialSource: string | undefined
-		try {
-			partialSource = fs.readFileSync(partialAbsolutePath, 'utf-8');
-		} catch (e) {
-			const fileWithError = path.join(genesisFileDirectory, currentFile)
-			console.error(`\x1b[31mPartial \x1b[1m${partialPath}\x1b[0m\x1b[31m does not exist\x1b[0m`)
-			console.error(`\t\x1b[2mError in ${fileWithError}\x1b[0m`)
+	private getCompiledPartials(): CompiledData[] {
+		const compiledPartials = [];
+		for (const [partial, source] of this.partials) {
+			const compiled = Handlebars.precompile(source, this.compileOptions);
+			const compiledData: CompiledData = [partial, compiled]
+			compiledPartials.push(compiledData);
 		}
-		return partialSource
+		return compiledPartials
 	}
 
-    static ImportScanner = class extends Handlebars.Visitor {
-        partials: Set<string>;
-        helpers: Set<string>;
-    
-        constructor() {
-            super();
-            this.partials = new Set<string>();
-            this.helpers = new Set<string>();
-        }
-    
-        PartialStatement(partial: hbs.AST.PartialStatement): void {
-            if (partial.name && partial.name.type === 'PathExpression') {
-                this.partials.add(partial.name.original);
-                return super.PartialStatement(partial);
-            } else {
-                throw new Error('Dynamic partial resolution is not supported');
-            }
-        }
-    }
+	private getTemplateSpecs(file: string): TemplateSpecification {
+		const extname = path.extname(file)
+		const name = path.basename(file);
+		const basename = path.basename(file, extname)
 
-	getCompiledPartial([partial, source]: SourceData) {
-		const compiled = Handlebars.precompile(source, this.handlebarsPluginOptions);
-        const compiledData: CompiledData = [partial, compiled]
-		return compiledData;
-	}
-
-	// Recursive function for getting nested partials with pathname
-	getPartialSources(name: string, source: string, genesisFileDirectory: string, list: SourceData[] = []) {
-		const filename = `${name}.hbs`
-		const relativeFileDirectory = path.dirname(name)
-
-		const tree = Handlebars.parse(source);
-		const scanner = new HandlebarsCompiler.ImportScanner();
-		scanner.accept(tree);
-
-		// Partials have been found
-		if (scanner.partials.size) {
-			for (const partialPath of scanner.partials) {
-
-				// Check if partial name is already registered in handlebarsPluginOptions
-				if (this.handlebarsPluginOptions.partials &&
-					typeof this.handlebarsPluginOptions.partials === 'object' &&
-					this.handlebarsPluginOptions.partials[partialPath]) {
-					continue
-				}
-
-				const partialSource = HandlebarsCompiler.getPartialSource(partialPath, genesisFileDirectory, filename)
-				// Skip if partial does not exist
-				if (!partialSource) continue
-				
-                const partialName = path.join(relativeFileDirectory, partialPath).replaceAll('\\', '/')
-                source = HandlebarsCompiler.renamePartial(source, partialPath, partialName)
-				list = this.getPartialSources(partialName, partialSource, genesisFileDirectory, list);
-			}
+        const compiledTemplateData = this.partials.find(([partial]) => {
+			return partial === basename
+		})
+		if (!compiledTemplateData) {
+			console.error('Error parsing template', file)
+			return
 		}
-        const sourceData: SourceData = [name, source]
-        list.push(sourceData)
-
-		return list;
-	}
-
-	getHelpers() {
-		let helpers: Helper[] = []
-		if (!this.handlebarsPluginOptions || !this.handlebarsPluginOptions.helpers || typeof this.handlebarsPluginOptions.helpers !== 'object') return helpers
-		helpers = Object.entries(this.handlebarsPluginOptions.helpers).filter(([helper, fn]) => typeof fn === 'function')
-		return helpers
-	}
-
-	getPartials() {
-		let partials: SourceData[] = []
-		if (!this.handlebarsPluginOptions || !this.handlebarsPluginOptions.partials || typeof this.handlebarsPluginOptions.partials !== 'object') return partials
-		partials = Object.entries(this.handlebarsPluginOptions.partials).filter(([partial, source]) => typeof source === 'string')
-		return partials
-	}
-
-	// Convert to ESM and register partial
-	toEsm(source: string, id: string): TemplateSpecification {
-		const dir = path.dirname(id);
-		const name = path.basename(id, '.hbs');
-
-		// Get nested partials
-		const partials = this.getPartials()
-		const partialsByPaths = this.getPartialSources(name, source, dir);
-		partials.push(...partialsByPaths)
-        const templateData = partials.find(([partial, source]) => partial === name)
-		
-		const children = [];
-		
-		for (const [partial, source] of partials) {
-			children.push(this.getCompiledPartial([partial, source]));
-		}
-		
-		const helpers = this.getHelpers()
+		const [templateName, compiledTemplate] = compiledTemplateData
 
 		// Create a tree
-		const tree = Handlebars.parse(templateData[1], { srcName: name });
-
-        const precompileOptions: PrecompileOptions = Object.assign({}, this.handlebarsPluginOptions)
+        const precompileOptions: PrecompileOptions = Object.assign({}, this.compileOptions)
         precompileOptions.srcName = name
-		const { code, map }: TemplateSpecification = Handlebars.precompile(tree, precompileOptions);
+		const tree = Handlebars.parse(compiledTemplate, { srcName: name });
+		const templateSpecs: TemplateSpecification = Handlebars.precompile(tree, precompileOptions);
 
-		let pluginTemplateData = {}
-		if (this.handlebarsPluginOptions && this.handlebarsPluginOptions.templateData && typeof this.handlebarsPluginOptions.templateData === 'object') {
-			pluginTemplateData = this.handlebarsPluginOptions.templateData
-		}
+        return templateSpecs
+	}
+
+	// Compile handlebars file to ESM
+	compile(file: string): CompileResult {
+		const compiledPartials = this.getCompiledPartials()
+		const helpers = this.helpers
+		const templateData = this.templateData
+
+		const { code, map } = this.getTemplateSpecs(file)
 
 		// Import this (partial) template and nested templates
 		const body = `
 			import Handlebars from 'handlebars/runtime.js';
 			const template = Handlebars.template(${code});
-			Handlebars.registerPartial('${name}', template);
 			${helpers.map(([helper, fn]) => `Handlebars.registerHelper('${helper}', ${fn});`).join('\n')}
-			${children.map(([partial, compiled]) => `Handlebars.registerPartial('${partial}', Handlebars.template(${compiled}));`).join('\n')}
+			${compiledPartials.map(([partial, compiled]) => `Handlebars.registerPartial('${partial}', Handlebars.template(${compiled}));`).join('\n')}
 			export default (data, options) => {
 				if (!data || typeof data !== 'object') {
 					data = {}
 				}
-				let templateData = Object.assign({}, ${JSON.stringify(pluginTemplateData)}, data)
+				let templateData = Object.assign({}, ${JSON.stringify(templateData)}, data)
 				return template(templateData, options)
 			};
 		`;
 
-		return {
-			code: body,
-			map,
-		};
+        return { code: body, map }
 	}
 }