+++ lib/index.ts
@@ -1,27 +1,23 @@
 import type { Plugin } from 'rollup';
-import HandlebarsCompiler from './handlebars-compiler';
-import { HandlebarsPluginOptions } from './types';
+import HandlebarsTransformer from './handlebars-transformer';
+
+import { HandlebarsPluginOptions } from './types/plugin-options';
 
 export default function handlebarsCompilerPlugin(handlebarsPluginOptions?: HandlebarsPluginOptions): Plugin {
 
-	const compiler = new HandlebarsCompiler(handlebarsPluginOptions)
+	const hbsTransformer = new HandlebarsTransformer(handlebarsPluginOptions)
 
 	return {
 		name: 'handlebars-compiler',
 
 		transform(source, id) {
 			if (/\.(hbs|handlebars)/.test(id)) {
-				// Get from cache when avalaible
-				if (compiler.cache.has(id)) {
-                    try {
-                        return JSON.parse(compiler.cache.get(id));
-                    } catch (e) {}
-				}
-
-				const output = compiler.toEsm(source, id);
-
-				compiler.cache.set(id, JSON.stringify(output));
-
+				const output = hbsTransformer.transform(source, id);
+				const existingWatchFiles = this.getWatchFiles()
+				const watchFiles = hbsTransformer.getWatchFiles(existingWatchFiles);
+				watchFiles.forEach((file: string) => {
+					this.addWatchFile(file)
+				})
 				return output;
 			}