+++ CHANGELOG.md
@@ -68,7 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [1.0.7](https://github.com/JohannIsaac/rollup-plugin-handlebars-compiler/pull/1/commits) 2020-8-25
+## [1.0.7](https://github.com/JohannIsaac/rollup-plugin-handlebars-compiler/pull/1) 2020-8-25
 
 ### Added
 
@@ -76,7 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [1.0.8](https://github.com/JohannIsaac/rollup-plugin-handlebars-compiler/pull/3/commits) 2020-8-25
+## [1.0.8](https://github.com/JohannIsaac/rollup-plugin-handlebars-compiler/pull/3) 2020-8-25
 
 ### Added
 
@@ -85,3 +85,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Cleaned up plugin source code into modules
+
+
+
+## [1.0.9](https://github.com/JohannIsaac/rollup-plugin-handlebars-compiler/pull/5)  2020-8-25
+
+### Changed
+
+- Source code refactor
+- Changed Rollup config to bundle types declaration file
+- Support for `--watch` flag for nested partials
+- Improved organization of types